@@ -1,27 +1,26 @@
 using System;
 using System.Linq;
 
-namespace xycat.common
+namespace xycat.common;
+
+public class Cipher
 {
-    public class Cipher
+    public static string Rot13(string value)
     {
-        public static string Rot13(string value)
-        {
-            var array = value
-                .ToCharArray()
-                .Select(c => c switch
-                {
-                    >= 'a' and <= 'z' when c  > 'm' => Convert.ToChar(c - 13),
-                    >= 'a' and <= 'z' when c <= 'm' => Convert.ToChar(c + 13),
+        var array = value
+            .ToCharArray()
+            .Select(c => c switch
+            {
+                >= 'a' and <= 'z' when c  > 'm' => Convert.ToChar(c - 13),
+                >= 'a' and <= 'z' when c <= 'm' => Convert.ToChar(c + 13),
 
-                    >= 'A' and <= 'Z' when c  > 'M' => Convert.ToChar(c - 13),
-                    >= 'A' and <= 'Z' when c <= 'M' => Convert.ToChar(c + 13),
+                >= 'A' and <= 'Z' when c  > 'M' => Convert.ToChar(c - 13),
+                >= 'A' and <= 'Z' when c <= 'M' => Convert.ToChar(c + 13),
 
-                    _ => Convert.ToChar(c)
-                })
-                .ToArray();
+                _ => Convert.ToChar(c)
+            })
+            .ToArray();
 
-            return new string(array);
-        }
+        return new string(array);
     }
 }
