@@ -1,13 +1,12 @@
 using System;
 
-namespace xycat.common
-{
-    public static class Constant
-    {
-        public static string StartLine = ">>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>";
-        public static string EndLine = "<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<";
+namespace xycat.common;
 
-        public static string StartChar = "[";
-        public static string EndChar = "]";
-    }
+public static class Constant
+{
+    public static string StartLine = ">>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>";
+    public static string EndLine = "<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<";
+
+    public static string StartChar = "[";
+    public static string EndChar = "]";
 }
